move code login to library crate, add github actions workflow to test and lint
